Refactor long files into smaller modules
