+++ tests/test_integration_processor.py
@@ -187,95 +187,3 @@ class TestChinaDataProcessorIntegration:
         hc_values = result.sort_values("year")["hc"].values
         assert all(hc_values[i] <= hc_values[i + 1] for i in range(len(hc_values) - 1))
 
-    @patch("builtins.open", new_callable=mock_open)
-    def test_markdown_output_format(self, mock_file, sample_raw_data):
-        """Test that markdown output is properly formatted."""
-        from utils.output import create_markdown_table
-
-        # Create sample processed data
-        processed_data = pd.DataFrame(
-            {
-                "Year": ["2020", "2021", "2022"],
-                "GDP": ["14722.73", "17744.64", "17886.33"],
-                "Population": ["1411.1", "1412.36", "1425.89"],
-            }
-        )
-
-        extrapolation_info = {"GDP_USD_bn": {"method": "ARIMA(1,1,1)", "years": [2023, 2024, 2025]}}
-
-        # Create markdown
-        create_markdown_table(processed_data, "test.md", extrapolation_info)
-
-        # Get written content
-        written_content = "".join(call.args[0] for call in mock_file().write.call_args_list)
-
-        # Check structure
-        assert "# Processed China Economic Data" in written_content
-        assert "| Year | GDP | Population |" in written_content
-        assert "## Data Sources" in written_content
-        assert "## Extrapolation to" in written_content
-
-    def test_error_handling_missing_columns(self):
-        """Test that processing handles missing columns gracefully."""
-        from utils.economic_indicators import calculate_economic_indicators
-
-        # Create data with missing columns
-        incomplete_data = pd.DataFrame(
-            {
-                "year": [2020, 2021],
-                "GDP_USD_bn": [14722.73, 17744.64],
-                # Missing many required columns
-            }
-        )
-
-        # Should not raise exception
-        result = calculate_economic_indicators(incomplete_data)
-
-        # Should return DataFrame
-        assert isinstance(result, pd.DataFrame)
-
-        # Should have original columns
-        assert "year" in result.columns
-        assert "GDP_USD_bn" in result.columns
-
-    @pytest.mark.parametrize("alpha,k_y_ratio", [(0.25, 2.5), (0.33, 3.0), (0.40, 3.5), (0.50, 4.0)])
-    def test_parameter_sensitivity(self, sample_raw_data, alpha, k_y_ratio):
-        """Test that different parameters produce different results."""
-        from utils.capital import calculate_capital_stock
-        from utils.economic_indicators import calculate_economic_indicators
-        from utils.processor_units import convert_units
-
-        # Process with different parameters
-        converted = convert_units(sample_raw_data)
-        capital_df = calculate_capital_stock(converted, capital_output_ratio=k_y_ratio)
-        converted["K_USD_bn"] = capital_df["K_USD_bn"]
-
-        result = calculate_economic_indicators(converted, alpha=alpha)
-
-        # TFP should vary with alpha
-        assert "TFP" in result.columns
-        assert result["TFP"].notna().any()
-
-    def test_csv_output_format(self, sample_raw_data):
-        """Test that CSV output is properly formatted."""
-        from utils.processor_dataframe.output_operations import prepare_final_dataframe
-
-        # Create column mapping
-        column_map = {"year": "Year", "GDP_USD_bn": "GDP", "POP_mn": "Population"}
-
-        # Convert units
-        from utils.processor_units import convert_units
-
-        converted = convert_units(sample_raw_data)
-
-        # Prepare final dataframe
-        result = prepare_final_dataframe(converted, column_map)
-
-        # Check column names
-        assert "Year" in result.columns
-        assert "GDP" in result.columns
-        assert "Population" in result.columns
-
-        # Check data integrity
-        assert len(result) == len(sample_raw_data)
-        assert result["Year"].dtype in [int, "int64", "int32"]
