+++ tests/test_wdi_downloader.py
@@ -165,61 +165,3 @@ class TestWDIDownloader:
         mock_sleep.assert_not_called()  # This should pass if download is successful on first try
 
     @patch("pandas_datareader.wb.WorldBankReader")
-    def test_download_wdi_data_column_naming(self, mock_wb_reader_class):
-        """Test that columns are named correctly."""
-        # Create data with dots in column name
-        data = pd.DataFrame({"year": [2020, 2021], "NY.GDP.MKTP.CD": [14722730.70, 17734062.65]})
-        mock_wb_reader_class.return_value.read.return_value = data
-
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-
-        # Column name should have dots replaced with underscores
-        assert "NY_GDP_MKTP_CD" in result.columns
-        assert "NY.GDP.MKTP.CD" not in result.columns
-
-    @patch("pandas_datareader.wb.WorldBankReader")
-    def test_download_wdi_data_year_column(self, mock_wb_reader_class):
-        """Test that year column is properly handled."""
-        data = pd.DataFrame({"year": [2020, 2021], "NY.GDP.MKTP.CD": [14722730.70, 17734062.65]})
-        mock_wb_reader_class.return_value.read.return_value = data
-
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-
-        # Year should be integer type
-        assert result["year"].dtype in [int, "int64", "int32"]
-
-        # Years should be in expected range
-        assert result["year"].min() >= 1960
-        assert result["year"].max() <= 2030
-
-    @patch("pandas_datareader.wb.WorldBankReader")
-    def test_download_wdi_data_data_types(self, mock_wb_reader_class):
-        """Test that data types are preserved correctly."""
-        data = pd.DataFrame({"year": [2020, 2021], "NY.GDP.MKTP.CD": [14722730.70, 17734062.65]})
-        mock_wb_reader_class.return_value.read.return_value = data
-
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-
-        # Check data types
-        assert result["year"].dtype in [int, "int64", "int32"]
-        assert result["NY_GDP_MKTP_CD"].dtype in [float, "float64"]
-
-    @patch("pandas_datareader.wb.WorldBankReader")
-    def test_download_wdi_data_return_type(self, mock_wb_reader_class):
-        """Test that function always returns a DataFrame."""
-        # Test with valid data
-        data = pd.DataFrame({"year": [2020], "NY.GDP.MKTP.CD": [1000]})
-        mock_wb_reader_class.return_value.read.return_value = data
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-        assert isinstance(result, pd.DataFrame)
-
-        # Test with exception
-        mock_wb_reader_class.return_value.read.side_effect = Exception("Error")
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-        assert isinstance(result, pd.DataFrame)
-
-        # Test with empty data
-        mock_wb_reader_class.return_value.read.side_effect = None
-        mock_wb_reader_class.return_value.read.return_value = pd.DataFrame()
-        result = download_wdi_data("NY.GDP.MKTP.CD")
-        assert isinstance(result, pd.DataFrame)
